perf(memory): stream parquet reads, write artifacts incrementally, store bulk arrays as f32
Peak resident set of the default single-run workflow is the maximum over stages, and
most stages were "materialise everything, then compute": `Table::read` decoded a whole
parquet file into `Vec<RecordBatch>` and the typed getters then copied each column into
an owned `Vec`, so both copies were alive during every load.

mumdia-io gains a streaming table layer: `TableFile` (footer only), `BatchReader`,
`for_each_batch`, typed column getters that decode batch by batch into one preallocated
output, and `TableWriter` for appending row groups. `Table::read` is gone from every
live stage, and `inspect` no longer decodes a multi-GB artifact to print ten rows.

Stage changes on top of that layer:

- extract streams chromatogram rows to parquet in 64k-row groups instead of holding
  every trace until one final `write_table`;
- convert accumulates 2048 spectra per row group instead of the whole run;
- the library fragment table is loaded by two streaming passes into preallocated flat
  arrays (the counting sort is unchanged), removing the Arrow transient that peaked at
  23 GB on a 657M-fragment library;
- quant collects the accepted candidate set first and skips non-accepted rows while
  streaming the chromatogram table;
- compete reads key columns and rewrites survivors batch by batch;
- the spectra loaders stream, and `Ms1Scan.mz` is f32 as the artifact stores it.

Bulk arrays move to f32 storage with f64 arithmetic:

- `FeatureMatrix` replaces the `Vec<Vec<f64>>` transpose in rescore with one flat
  row-major f32 buffer sized from the parquet footers. The standardizer accumulates
  means and variances in f64, `std_row` divides in f64, and the logistic gradient and
  score widen per term. Standardised per-fold training copies are f32 too, which is
  where `native_tda` spends `folds` times the matrix. At 11.6M PSMs by 387 features
  that is 33.6 -> 16.8 GiB for the matrix and roughly 135 -> 67 GiB for the
  `native_tda` peak.
- `Library.frag_mz` is f32: 26 -> 22 bytes per fragment row. Both matchers already
  quantised to f32 before use (fragindex bins and stores `mz as f32`, naive compared
  `mz as f32 as f64`), so matching is bit-identical; consumers that want f64 widen at
  the call site.

Layout changes are value-preserving. The f32 storage changes values in the seventh
significant digit and must clear the acceptance gates in docs/27 section 5 (peptide
count within 0.3%, empirical decoy fraction unchanged, Jaccard of accepted
peptidoforms at least 0.995, quant bias and CV within 0.005) before it is treated as
validated.

+++ rust/mumdia/crates/mumdia/src/matchers/naive.rs
@@ -27,7 +27,7 @@ pub fn score_scan_count_dot(
         let mut dot = 0.0f64;
         for k in 0..cand.n_frag {
             let gi = cand.frag_start + k;
-            let fmz = lib.frag_mz[gi] as f32 as f64; // match FragIndex f32 storage
+            let fmz = lib.frag_mz[gi] as f64; // library stores f32, as FragIndex matches on
             let fint = lib.frag_int[gi];
             for &(pmz, pint) in peaks {
                 if within_ppm(fmz, pmz, tol_ppm) {
